requirements.txt
spacy
transformers
torch 
torchvision
fairseq
sacremoses
flask 
python-dotenv
google-cloud-translate